# docs/swagger.yaml
basePath: /api/v1
definitions:
  subscriptions.ErrorResponse:
    properties:
      error:
        example: invalid request body
        type: string
    type: object
  subscriptions.ListResult:
    properties:
      has_next:
        example: true
        type: boolean
      has_prev:
        example: false
        type: boolean
      next_page:
        example: 2
        type: integer
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      prev_page:
        example: 0
        type: integer
      subscriptions:
        items:
          $ref: '#/definitions/subscriptions.Subscription'
        type: array
      total:
        example: 25
        type: integer
      total_pages:
        example: 3
        type: integer
    type: object
  subscriptions.Subscription:
    properties:
      end_date:
        example: 12-2024
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        type: string
      price:
        example: 999
        type: integer
      service_name:
        example: Netflix
        type: string
      start_date:
        example: 01-2024
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        format: uuid
        type: string
    type: object
  subscriptions.SumResult:
    properties:
      total:
        example: 1998
        type: integer
    type: object
host: localhost:8800
info:
  contact: {}
  description: API для управления подписками.
  title: Subscriptions API
  version: "1.0"
paths:
  /sub:
    get:
      consumes:
      - application/json
      description: Возвращает список подписок с фильтрацией, сортировкой и пагинацией.
      parameters:
      - description: Фильтр по названию сервиса
        in: query
        name: service_name
        type: string
      - description: Алиас для service_name
        in: query
        name: q
        type: string
      - description: UUID пользователя
        in: query
        name: user_id
        type: string
      - description: Поле сортировки
        enum:
        - service_name
        - price
        - start_date
        - end_date
        in: query
        name: sort
        type: string
      - description: Направление сортировки
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - default: 1
        description: Номер страницы
        in: query
        minimum: 1
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscriptions.ListResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/subscriptions.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/subscriptions.ErrorResponse'
      summary: Получить список подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Создаёт новую подписку. Если id не передан, он генерируется автоматически.
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/subscriptions.Subscription'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/subscriptions.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/subscriptions.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/subscriptions.ErrorResponse'
      summary: Создать подписку
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновляет подписку по id из тела запроса.
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/subscriptions.Subscription'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/subscriptions.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/subscriptions.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/subscriptions.ErrorResponse'
      summary: Обновить подписку
      tags:
      - subscriptions
  /sub/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет подписку по UUID.
      parameters:
      - description: UUID подписки
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/subscriptions.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/subscriptions.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/subscriptions.ErrorResponse'
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      consumes:
      - application/json
      description: Возвращает одну подписку по UUID.
      parameters:
      - description: UUID подписки
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscriptions.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/subscriptions.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/subscriptions.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/subscriptions.ErrorResponse'
      summary: Получить подписку по ID
      tags:
      - subscriptions
  /sub/sum:
    get:
      consumes:
      - application/json
      description: Возвращает суммарную стоимость подписок за период с optional-фильтрами
        по пользователю и сервису.
      parameters:
      - description: Начало периода в формате MM-YYYY
        in: query
        name: start_date
        required: true
        type: string
      - description: Конец периода в формате MM-YYYY
        in: query
        name: end_date
        required: true
        type: string
      - description: Фильтр по названию сервиса
        in: query
        name: service_name
        type: string
      - description: Алиас для service_name
        in: query
        name: q
        type: string
      - description: UUID пользователя
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscriptions.SumResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/subscriptions.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/subscriptions.ErrorResponse'
      summary: Посчитать сумму подписок
      tags:
      - subscriptions
schemes:
- http
swagger: "2.0"